+++ src/days/day_21.rs
@@ -22,9 +22,9 @@ pub fn run() {
 	// J = D & (!A | !B | !C) & (H | E)
 	for command in commands {
 		for c in command.chars() {
-			code.inputs.push(c as u8 as i64);
+			code.inputs.push(c as u8 as isize);
 		}
-		code.inputs.push(b'\n' as i64);
+		code.inputs.push(b'\n' as isize);
 	}
 	while let Some(c) = int_code(&mut code, true) {
 		if c > 255 {
@@ -47,9 +47,9 @@ pub fn part_one() {
 	// J = D & (!A | !B | !C)
 	for command in commands {
 		for c in command.chars() {
-			code.inputs.push(c as u8 as i64);
+			code.inputs.push(c as u8 as isize);
 		}
-		code.inputs.push(b'\n' as i64);
+		code.inputs.push(b'\n' as isize);
 	}
 	while let Some(c) = int_code(&mut code, true) {
 		if c > 255 {
